Added parameter instantiation + generic interface class + factory
For each derived `BoundaryValues` class, added a struct to the header file which
holds relevant parameters, and also created a corresponding constructor to read
the parameters.
Also added an interface class `BoundaryValuesInterface` which does not include
the dimension, and from which `BoundaryValues` is derived.
This allows us to create a factory which can give back any `BoundaryValues`
derived class, and also which will give back the class with the proper dimension.
The parameter structs are derived from a `BoundaryValuesParams` struct which just
contains a name -- this is so that the factor can read in that struct type in
order to accept all of the parameter structs.