Wrote `EvaluateFEObject` to print systems at (grid)points.
Wrote class `EvaluateFEObject` which reads off points or a grid from an hdf5
file, evaluates a finite element solution at those points (given a DoFHandler
and solution vector) and then writes it to a new hdf5 file, either as a grid if
it read in a grid, or as a list of points if that's what it read off.

Additionally, wrote function in `IsoSteadyState` class which uses
`EvaluateFEObject`, and called that function in `IsoSteadyStateSim` (along with
appropriate setup and command-line argumens).

Since we introduced a new class, had to modify several CMakeLists.txt files in
order to link everything together properly.

Also, changed Python files to accomodate these new written hdf5 files.

Finally, completely butchered the `read-defect-print-to-grid.cpp` file to grab
what turned into `EvaluateFEObject`.
This file will eventually write the FE object to a grid by itself (using
`EvaluateFEObject`), but we first need to figure out how to properly archive an
`IsoSteadyState` object.