Added minimal example to demonstrate dealii ghosted vector behavior
This minimal example creates a ghosted vector and a completely distributed
vector, assigns values to ghosted vector from completely distributed vector, and
prints out the result. Then, it adds the values from the completely distributed
vector which (as discussed in the dealii documentation on ghosted vectors) does
not update the ghost values, resulting in presumably unwanted behavior.