Copied in deal.II step-40 with code that causes library bug
There is a bug (heretofore unfound) which causes executables to be linked to
PETSc without the proper global path. This causes a runtime linking error unless
the LD_LIBRARY_PATH has the PETSc shared libary in it. As a demonstration of
this bug, I modified the existing step-40 dealii program to update the global
system matrix and right-hand-side separately. See line 421 for the additions
which cause the bug.