Messed with boundary conditions for 3D.
Added a little section in grid generation which marks the top and bottom of the
hypercube as not part of the Dirichlet boundary (this is only for 3D). The
problem gets weird stripes at the domain boundaries for the parallel
implementation, so it is commented out for the time being.