Added additional command-line options to IsoSteadyStateMPI
Added ability to refine different amounts along different axes. Also added
ability to set the dimension at run-time, and to set whether or not the solver
used the AMG preconditioner. Finally, made the simulation output the number of
DoFs as well as the dimension (just to clear up potential confusion later).