Added BoundaryValues class which implements two defects
TwoDefectConfiguration class is just a BoundaryValues object which creates two
defects. So far it only has two configurations, `plus_half_minus_half` and
`plus_half_minus_half_alt`, the latter of which corresponds to an isomorphism of
director configurations. The latter is also not implemented (indeed, choosing
that configuration just gives you the other at this point)