Made solution vector a block vector to hold Q-vector
In order to use a Q-vector field as a source, I need to structure it as
something which is handled by the same DoFHandler object. This way we can index
the quadrature values of the Q-tensor using the same cell iterator. The only way
I could think to do this is to structure it as a BlockVector object which only
couples to itself (and not the velocity or pressure). This works, but it was a
huge pain to implement because of difficult-to-understand errors (e.g. seg
faults, bad linked lists, etc.). This commit gets it to work, however.