Added a utility to read in simulation parameters from command-line
Wrote a little function read_command_line_options in the namespace
SimulationOptions which basically just takes in the (ac, av) objects from main
and spits back out a program_options::variables_map object. All I did was copied
and pasted from the main IsoTimeDependentSim. This gives you basically a
one-liner to read in the copious number of command-line arguments for the
simulation. Used it in IsoTimeDependentSim and it seems to work. Still need to
figure out how to organize it so something is useful for *all* the simulations.