Added ability to read in Q-tensor config to hydro simulation
HydroFixedConfiguration only computes the Stokes equation solution given a
Q-tensor configuration. To this end, we need to read in a Q-tensor configuration
that has already been computed. We do this by creating an FEFieldFunction object
from the IsoTimeDependent object. Then we pass that into a function which works
with the BlockVector structure (just calls the FEFieldFunction object for
components which correspond to Q, gives zero otherwise). This function can then
project onto a temporary BlockVector and then we can copy over the Q-tensor
block to the actual Q-tensor solution in the hydro BlockVector. Note that we
also have outputted each of the components from both of the Q-configuration
representations, just to make sure the projection operation worked properly.