Made the distortion stress tensor the source term
Up until now I had just used the analytic expression for the forcing term due to
active stresses from a defect (from Giomi). Here I read in a Q-tensor
configuration, and then compute the distortion stress tensor for the forcing
term. Still need to add in the contribution from the variation of the free energy.