Fixed bug where A_preconditioner was factorized
The `factorize` method is not available with the SparseILU preconditioner. I
guess I had added it in to see if something was going wrong, but then never took
it out because everything worked fine in 2D. Got brought up because
IsoTimeDependentHydro complained when it tried to compile with this error.