Added code to compute reduced LagrangeMultiplier
The Lagrange multiplier typically has 5 degrees of freedom. However, since it is
diagonalized by the same matrix that Q is, it can be represented by two degrees
of freedom as well as a rotation matrix (which can be computed directly from Q).
This program computes the Lagrange multiplier more efficiently by only computing
those two degrees of freedom. The diagonalization process (as well as the
automatic differentiation of those values) is done externally.