Wrote code for more efficient Lagrange multiplier solver
Came up with a scheme to first diagonalize Lambda-tensor before doing the
Newton's method, which reduces the number of integrals which must be computed by
a factor of 10 or so. Indeed, the timing example shows that (on my machine) the
computation time is reduced by a factor of 20. In future iterations I hope to
reduce this by another factor of 8 by exploiting the symmetry of the spherical integrals.