Used symmetry to reduce quadrature points for Lebedev integrals
For Lebedev quadrature, each of the points is generated by some initial point
which is rotated around the sphere by the action of some element in the
octahedral group. For points that look like (0, 0, 1) etc. there are six other
corresponding points, three of which have non-negative x, y, and z components. For
points that look like (0, A, A) for A = 1/sqrt(2), there are twelve other
corresponding points, three of which have non-negative x, y, and z components.
For the rest, they exist on the positive octant of the sphere and correspond to
8 other points, one in each octant (although the element of the octahedral group
typically generates many more points in the positive octant from the given
point). In any case, because all of our integrals only contain x^2 and y^2, we
may get rid of all of the corresponding points outside the positive octant, and
just multiply the weights by the number of other corresponding points (2, 4, and
8 respectively for the cases mentioned above).