Wrote examples to refine grid like zumer, check Lagrange continuity
Refined the grid only at the points where the defects were to try to get a
little bit more resolution in the same way that Zumer did.
Additionally, wrote a program which outputs the singular potential and
Jacobian-of-singular potential values out for a system with perturbed degenerate
eigenvalues as the perturbation goes to zero and becomes negative.
It happens that the function and its derivative are both continuous in that
case, so presumably we can write down a continuous analytic expression.