Stripped down HydroFixedConfiguration
In an effort to modularize the code, I stripped down HydroFixedConfiguration to
its barebones.
That means I got rid of any reference to IsoTimeDependent because (for sake of
modularity) HydroFixedConfiguration should not have to deal with the
Q-configuration.
Instead, for now it just reads from a function which gives back the values
calculated analytically from a +1/2 active defect.
Additionally I made it so that it reads in a triangulation, because that should
exist outside of the hydro class (that should be handled by some other
simulation object).