Turned HydroFixedConfiguration into class rather than executable
Before, HydroFixedConfiguration was a sort of hacky class which ran
IsoTimeDependent for a few steps, and then calculated the flow configuration.
This was not super modular and presented problems with testing and extending.
Now, HydroFixedConfiguration is a class which only handles setting up the finite
element function corresponding to flow, and solves the system.
Actually dealing with the forcing function, setting up the domain, etc. (things
which can be shared by nematics and flow) are handled by the simulation drivers.