Made HydroFixedConfiguration take stress tensor, rather than force
Up until this point we have inputted the right-hand side of the hydro (Stokes)
equation by just giving back a force f.
Instead, we now input a stress tensor.
This manifests in the weak form of the equations by integrating the right-hand
side by parts so we take the inner product of the stress tensor with the
gradient of the test velocity.
In code, this means we have to pass in a (unique pointer to a)
dealii::TensorFunction object.