Added function to solve entire block of HydroFixedConfiguration
The standard way of solving Stoke's equation is by doing this Schur complement
thing -- this involves working with the individual blocks of the BlockMatrix.
Instead, because we will have to *not* work with the individual blocks, we try
to run a CG solver on the whole thing.
This turns out to work fine (I thought it would give me compiler problems).