Made `HydroFixedConfiguration` compatible with mu_2 term
Added the contributions from the mu_2 viscous term in the viscous stress tensor
to the HydroFixedConfiguration solver.
This included being able to pass in a Q-tensor configuration as a `TensorFunction`
via the `assemble_system` function.
Additionally, have to use a direct solver because GMRES doesn't converge and I'm
not sure that I can trust CG for a nonsymmetric system.