Added assemble_system function which assembles hydro system
I will need to couple together the hydrodynamic and Q-tensor systems, and so the
simulation drivers will need to be able to assemble the finite element systems.
Hence, I have written one that just assembles the system for the hydro by
passing in the hydro object and then internally calling the return functions so
that we may use dof_handler, fe_system, constraints, and modify system_matrix,
preconditioner_matrix, system_rhs.
This seems to work well, except it is a little dicey as far as encapsulation is concerned.