Only keep track of one past configuration, rather than a vector
In this simulation it was the case that we would keep track of all past
configurations in a `std::vector<dealii::Vector>` object.
Obviously this is extremely memory-inefficient, and puts a limit on how many
time-steps we can do.
Additionally, it cluttered up many functions by having to reference the previous
time-step.
Instead what we will do is just outpute the past_timestep vector if we want to
save previous configurations -- that way they go to disk and we don't have to
worry about consuming too much memory.