Rearranged the unused functions to be at the end
There are several functions which I used for debugging purposes or some such
other nonsense, and I just put them at the end.
Ideally I would delete them and do something else with the example or simulation
programs that use them, but that's tedious and not a priority right now.