Wrote new constructors for `BoundaryValues` objects.
Before, the only way to construct `BoundaryValues` objects was to explicitly
input arguments (doesn't work for polymorphism) or input a boost::variables_map
object (which doesn't let you set the arguments outside of the command line
rigamarole).
To ameliorate this, I wrote a constructor which takes in a std::map<std::string,
boost::any> object so that we can input the parameters and still use it with
polymorphism.