Changed LagrangeMultiplier so coords/weights are allocated in 1 step
Previously, the Lebedev coords and Lebedev weights were static attributes of the
LagrangeMultiplier class which had to be allocated in the source file via static
methods.
This was weird and clunky (and also required a template argument) so now we use
a regular class method to create the Lebedev coords and weights in one go.
These are stored in a class LebedevCoords which lives in the LagrangeMultiplier
class.