Wrote more modular liquid crystal simulation
Previously, programs like IsoSteadyState, IsoTimeDependent, IsoSteadyStateMPI,
etc. have done all of the read-in, grid setup, and finite element solving in one
object.
I am trying to make that process more modular so that a SimulationDriver object
handles all of the overhead particular to a given simulation, and then the
LiquidCrystalSystem (or HydroSystem) handles the actual configuration, solving,
assembling matrices, whatever.
This way it is easy to swap out different pieces, and also to couple the
different finite element systems.