Added serialization to LiquidCrystalSystem class, tested out
It's useful to serialize C++ objects so that you can just read them in and pick
back up where you were.
I added serialization to the LiquidCrystalSystem class so that I didn't have to
wait for the system to relax every time I wanted to try solving a hydro
configuration on it.