Broke `setup_system` into two different functions
Broke `setup_system` into `setup_dofs` which sets up the dof_handler, the
affine_constraints, the sparsity pattern, and then initializes the system_matrix
and rhs; and `initialize_fe_field` which initializes the current and past
solution vectors to whatever the BoundaryValues function is.
I did this to help with parallel serialization -- essentially the only things we
need to serialize are the triangulation and the current solution vector.
Everything else can be done rather quickly once we have those two pieces of information.