Wrote code to serialize NematicSystemMPI
Wrote some code to serialize all of the data associated with the
NematicSystemMPI.
The NematicSystemMPIDriver takes care of most of it.
Additionally, had to use a SolutionTransfer object to serialize the distributed
triangulation and solution vector object.

One thing to note is that it doesn't seem to work if you run on different
numbers of processors, unfortunately.
Need to figure that out.