Fixed different-mpi-process serialization issue
Had a bug where serialization would fail when you serialize and deserialize on
different numbers of mpi processes.
This was a bug due to me serializing the dof_handler object when really I just
need to deserialize the corresponding triangulation, make sure the FESystem
object is generated correctly, and then just regenerate the dof_handler.