Fixed bug with hydrocoupler that messed with rhs integral
Had a bug where I had put the parenthesis in the wrong place in the rhs
expression for calculating hydro.
This caused the dx expression (quadrature weight) to only be multiplied by the
last term in the expression, thus making the flows corresponding to all the
other terms really big.