Added ability to make progressively finer grid for nematic
Wanted to more closely match Zumer hydrodynamic setup, so I added a method which
generates a gride that gets finer and finer as you get closer and closer to the
central portion of the domain.
This allowed me to have a discretization which was slightly finer than Zumers
while having enough memory to run on my computer.