Added in defect tracking to nematic simulation
Already wrote program to track defects, now I have included it in the
actual simulation and written a little method to output the defect data
in parallel.
Still have some kind of a bug that (I think) happens when no defects are
found at the time of writing the hdf5 file.
Need to figure out where that is and fix it.