Added autodiff anisotropic elasticity
Trying to figure out how to implement anisotropic elasticity into the
very basic nematic system.
However, things aren't working out immediately so I implemented the
residuals as given (since they match up with the periodic configuration)
and then added autodifferentiation in order to calculate the
corresponding Jacobians.
This actually works, and so the plan is to compare the analytically
calculated Jacobians to the autodiff Jacobians to try to locate the
problem.