Changed defect shape to have logistic S
Was having some issues with convergence of anisotropic elastic terms
(had to make time-step really small), so at Cody's suggestion I made the
Q-tensor a smooth function.
This involved choosing S to be a function which goes to 0 at the core.
I chose a logistic function because it looks to me like the final
S-value for a relaxed isotropic elasticity defect.