Fixed minor bug with defect identification
The bug is that, in order to find defects we calculate the maximal
charge on each cell in a domain, and then make sure that the absolute
value is above some threshold (that gives a defect).
However, for negatively charged defects we should actually be looking
for the minimum value (because the minimum value will have the largest
magnitude).
This hasn't been a huge deal because the maximum on those cells is still
quite negative, but it was worthwhile to fix, if only to add another
field called `min_D` to the defect properties struct.
This allows us to write the charge associated with a given defect.