Outputted charge from `find_defects`
Before I was only outputting position from `find_defects` function, but
now I've outputted it as a tuple.
Also outputted that from the nematic simulation so that it gets written
to the hdf5 file (this will make it way easier to distinguish between
the two defects in the annihilation analysis).