Fixed bug with L2 Jacobian, have weird behavior
I had a typo in the sympy code that generates the expressions for the
C++ code in generating the L2 Jacobian.
Fixed that, but this weird thing is still happening where I get ripples
in the nematic configuration and then the Newton steps stop converging.
Still have to figure that out.