Made Dzyaloshinskii parameters class attribtutes
Have a whole bunch of parameters corresponding to the
DzyaloshinskiiSystem which are no longer class attributes of
DzyaloshinskiiSystem, but rather which are passed in as class method
parameters.
Here we just made them attributes of the DzyaloshinskiiFunction class
adn then had to change the constructors and initialize function to match.