Added more functions for parsing
Added ability to parse a list of numbers that are separated by commas.
This is essentially the same as the other parsing by delimited, except
it returns doubles.
Additionally, wrote some functions to translate between vector of
deal.II points and vector of vectors (this is necessary for some of the
        MPI functions).