Fixed bug in non-defect-refined simulations
In the code that handles extra coarsening and refinement around the
mesh, there was a piece which assumed that there was at least one
further refine around defects, and dereferenced the minimal further
refine.
In the case that there were no further refines, that caused a seg fault.
Here I just fixed it by adding a check that there were actually further
refines.