Added functionality to freeze defects
To freeze defects, one has to fix the degrees of freedom whose support
points live on the edges of cells which form a boundary between regions.
These regions can be defined in any way, but we define them as a region
of fixed radius around each defect, and the complement of such regions.
To fix these support points, they're added as algebraic (affine)
constraints to the corresponding `AffineConstraints` object which also
handles Dirichlet boundary conditions and hanging node constraints.
Functions are added to the NematicSystem functions to make this work,
          and the defects can be frozen with a boolean parameter in a
          parameter file.