Notes on how to restructure simulations
As it stands, the whole codebase is spaghetti.
Taking inspiration from deal.II step-13, I'm hoping to rewrite most of
the codebase so that it is more modular, and therefore easier to
maintain.
These notes take down ideas from step-13 and also try to explain
succinctly what functionality is provided currently so that I can sort
and recreate that functionality in a more modular way.
Eventually will also need to include some notes about issues that I
filed in GitHub.