Added function to make parallel dofs consistent
In the internal boundary value constraints program I'm running into this
problem where, because of how domains are split up, we can get
inconsistencies at dofs which are on the edges of domains.
Apparently there's a deal.II function for dealing with this in the
AffineConstraints object, so we implemented that.
We'll see if it works, I suppose.