field initialization and boundary values decoupled
In order to freeze the defects in place, I need the ability to specify
the boundary values (i.e. values at the edges and around the defects)
    separately from the initialization of the field (which I wanted to
            be uniaxial two-defect configuration in isotropic
            elasticity).
This is because the other scheme (just melting everything away from the
        defects) caused there to be a creation of domains and defects
when the rest of the configuration froze.
It's sort of ad hoc right now, and will need to update it once I
refactor the whole codebase.