Separated semi-implicit assembly
Trying to modularize NematicSystemMPI, and the first step for that is
taking all of the class methods out and separating them into their own
libraries.
To start, I just took out the semi_implicit singular potential matrix
assembly method.
Seems to work, so I'm goingn to do that with the rest of the assembly
methods.