Can get function values at radial points
In several scenarios, it's necessary to query a finite element field
at many points along some number of concentric circles around a given
point (e.g. investigating the Fourier modes around a point as a function
        of radial distance from the point).
To facilitate this, I wrote a function which internally generates the
points at which to query based on a `RadialPointSet` struct, and then
calls the relevant functions to query the field at those points.