PerturbativeDirectorSystem runs correct simulation
Put together this simulation to compare Carter's linearized equation
numerical solution to the full Ball-Majumdar simulation.
To this end, I set the parameters to be the same system as my full
Ball-Majumdar simulation from before.
Thus, running this should give something that can be compared to the
full Q-tensor Ball-Majumdar simulation.