Now fix internal boundaries with AffineConstraints
Have been setting internal boundaries with a map of global dof indices.
This is not great because then I have to manually put it into an
AffineConstraints object.
Additionally, have to make the AffineConstraints object consistent in
parallel.
Hence, I just put that into its own function which internally calls the
earlier function.