Specialized defect functions to throw errors in 3D
There are two functions `find_defects` and `output_defect_positions`
which I'm unsure how to generalize to 3D.
To make a 3D system work ad hoc, I just specialized them and threw an
exception in the 3D case.
Additionally, wherever I used them I put a check for dimensionality so
that they're not called in the 3D case.