Fixed bug with grid naming conventions.
Switched over to the built-in deal.II function to generate a grid from
name and arguments to increase flexibility for the user.
However, I had been calling it "hypercube" and deal.II calls it
"hyper_cube", so in the `refine_further` function it bipassed and just
refined all cells.
I fixed this by continuing the next loop iteration if the grid name was
not one of "hyper_cube" or "hyper_ball", and also made all the naming
conventions match with deal.II's.