Now input boundary values as an array of tables with boundary_id
Previously one could only specify a single boundary_values object which
would be applied to boundary_id = 0.
Now one may specify any number of boundary conditions by feeding an
array of tables which is a boundary_values table and also includes a
boundary_id to specify which boundary is being referred to.