Fixed LagrangeMultiplierReduced to it returns correct Z value
Prior to this change, the Z-value was off by a factor of 4pi
(due to some details with the Lebedev quadrature).
This doesn't actually change the dynamics (since this comes out as a
constant in the actual energy), but it will change its value.
I have corrected this so that the energy is exactly given by the stated
expression.